Break out logic to handle archived until escalating
